bus renamed to scheduler
