+++ scheduler/pom.xml
@@ -3,12 +3,12 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>bus</artifactId>
+    <artifactId>scheduler</artifactId>
     <version>0.0.1-SNAPSHOT</version>
     <packaging>jar</packaging>
 
-    <name>bus</name>
-    <description>Bus part of blackbook project</description>
+    <name>scheduler</name>
+    <description>Scheduler part of blackbook project</description>
 
     <parent>
         <groupId>com.blackbook</groupId>
