+++ scheduler/src/main/java/com/blackbook/scheduler/controller/core/RequestControllerListener.java
@@ -1,4 +1,4 @@
-package com.blackbook.bus.controller.core;
+package com.blackbook.scheduler.controller.core;
 
 /**
  * @author Siarhei Shauchenka
