+++ scheduler/src/main/java/com/blackbook/scheduler/model/ObserverRepository.java
@@ -1,4 +1,4 @@
-package com.blackbook.bus.model;
+package com.blackbook.scheduler.model;
 
 import org.springframework.data.repository.CrudRepository;
