+++ scheduler/src/main/java/com/blackbook/scheduler/processor/StartAllCrawlersProcessor.java
@@ -1,12 +1,7 @@
-package com.blackbook.bus.processor;
+package com.blackbook.scheduler.processor;
 
-import com.blackbook.bus.controller.core.RequestControllerListener;
-import com.blackbook.bus.processor.core.AbstractProcessor;
-import com.blackbook.bus.processor.core.RequestProcessor;
-import com.mashape.unirest.http.HttpResponse;
-import com.mashape.unirest.http.JsonNode;
-import com.mashape.unirest.http.Unirest;
-import com.mashape.unirest.http.exceptions.UnirestException;
+import com.blackbook.scheduler.controller.core.RequestControllerListener;
+import com.blackbook.scheduler.processor.core.AbstractProcessor;
 
 /**
  * @author Siarhei Shauchenka
