+++ scheduler/src/main/java/com/blackbook/scheduler/processor/core/RequestProcessor.java
@@ -1,4 +1,4 @@
-package com.blackbook.bus.processor.core;
+package com.blackbook.scheduler.processor.core;
 
 /**
  * @author Siarhei Shauchenka
