+++ scheduler/src/main/resources/migration/V1_01__initial.sql
