+++ scheduler/src/test/java/com/blackbook/scheduler/SchedulerApplicationTests.java
@@ -1,4 +1,4 @@
-package com.blackbook.bus;
+package com.blackbook.scheduler;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -7,7 +7,7 @@ import org.springframework.test.context.junit4.SpringRunner;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest
-public class BusApplicationTests {
+public class SchedulerApplicationTests {
 
 	@Test
 	public void contextLoads() {
